% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridrho.f.R
\name{gridrho.f}
\alias{gridrho.f}
\title{Support function for \code{\link{ui.causal}}}
\usage{
gridrho.f(rho, gridn, rho.plotrange, plot)
}
\arguments{
\item{rho}{interval that should be divided}

\item{gridn}{number of gridpoints}

\item{rho.plotrange}{a larger interval of grids to be used in a plot}

\item{plot}{whether or not the larger interval of grids should be created}
}
\description{
Divides the \code{rho} interval into a grid
}
