% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grr.R
\name{grr}
\alias{grr}
\title{Gradient for the loglikelihood used by ui.probit}
\usage{
grr(par, rho, X.z = X.z, X.y = X.y, y = y, z = z)
}
\arguments{
\item{par}{Coefficients.}

\item{rho}{Rho.}

\item{X.z}{Covariate matrix for missingness.}

\item{X.y}{Covariate matrix for outcome.}

\item{y}{Outcome.}

\item{z}{Missing or not.}
}
\description{
This function derives the gradient in order for \code{\link{ui.probit}} to run faster.
}
