% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interv.p.R
\name{interv.p}
\alias{interv.p}
\title{Print function for intervals}
\usage{
interv.p(v, digits = 3)
}
\arguments{
\item{v}{Lower and upper bounds.}

\item{digits}{Number of decimals.}
}
\description{
This is a support function for \code{\link{print.ui}} which allows you to print an interval (vector of two elements) using parenthesis.
}
