% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda0.R
\name{lambda0}
\alias{lambda0}
\title{Inverse mills ratio d(x)/(1-p(x)) #' Inverse mills ratio d(x)/p(x) which is the standard normal density function divided by one minus the standard normal cumulative distribution}
\usage{
lambda0(x)
}
\arguments{
\item{x}{Vector}
}
\description{
This is a support function for \code{\link{ui.y0}}, \code{\link{ui.y0t1}} and \code{\link{ui.y1t0}} that allows you to calculate the inverse Mills ratio d(x)/(1-p(x)).
}
