% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogL.probit.R
\name{LogL.probit}
\alias{LogL.probit}
\title{Loglikelihood used by ui.probit}
\usage{
LogL.probit(par, rho, X.z = X.z, X.y = X.y, y = y, z = z)
}
\arguments{
\item{par}{Coeficient values the logliklihood should be drived at.}

\item{rho}{The value of the sensitivity parameter.}

\item{X.z}{covariate matrix for missingness mechanism}

\item{X.y}{covariate matrix for the outcome regression}

\item{y}{outcome vector}

\item{z}{indicator of wether y is missing or not}
}
\description{
This function derives the Loglikelihood for \code{\link{ui.probit}}.
}
