% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logl.sandACT.R
\name{Logl.sandACT}
\alias{Logl.sandACT}
\title{Loglikelohood used in sandwich estimator of average causal effect on the treated for DR}
\usage{
Logl.sandACT(x, X, z)
}
\arguments{
\item{x}{coefficents.}

\item{X}{Covariate matrix.}

\item{z}{Missing or not.}
}
\description{
Loglikelohood used in sandwich estimator of average causal effect on the treated for DR, support function for \code{\link{ui.causal}}
}
