% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.probit.R
\name{ML.probit}
\alias{ML.probit}
\title{Fit maximum likelihood for fixed values of rho}
\usage{
ML.probit(
  out.formula,
  mis.formula = NULL,
  data,
  rho = c(-0.5, 0.5),
  progress = TRUE,
  method = "NR"
)
}
\arguments{
\item{out.formula}{Formula for outcome regression.}

\item{mis.formula}{Formula for regression model for the missingness mechanism.}

\item{data}{Data frame containing the variables in the formulas}

\item{rho}{Vector containing the values of rho for which we want to fit the likelihood.}

\item{progress}{If TRUE prints out process time for each maximazation of the likelihood.}

\item{method}{Maximazation method to be passed through \code{maxLik}}
}
\description{
This is a support function for \code{\link{ui.probit}}
}
