% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.uicausal.R
\name{plot.uicausal}
\alias{plot.uicausal}
\title{Plot of UI and CI}
\usage{
\method{plot}{uicausal}(x, DR = TRUE, main = "", xlab = NULL, ylab = "", ...)
}
\arguments{
\item{x}{An object of class uicausal}

\item{DR}{If TRUE the doubly robust estimator is plotted, otherwise the outcome regression estimator is plotted.}

\item{main}{Main title, default is no title.}

\item{xlab}{Title for xaxis, default is \code{expression(rho)}.}

\item{ylab}{Title for y axis, default is no title.}

\item{...}{Additional arguments, use is discouraged.}
}
\description{
Plot function for objects returned from \code{\link{ui.causal}}.
Plots confidence intervals for different values of rho and the uncertainty interval.
}
