% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.uipcor.R
\name{plot.uipcor}
\alias{plot.uipcor}
\title{Plot of UI and CI for partial correlation}
\usage{
\method{plot}{uipcor}(
  x,
  plot.all = TRUE,
  which = NA,
  intercept = FALSE,
  ylab = NULL,
  col = c("black", "red"),
  ...
)
}
\arguments{
\item{x}{An object of class uiols}

\item{col}{Vector containing the color of confidence intervals (default black) and uncertainty intervals (default red).}

\item{...}{Additional arguments, use is discouraged.}
}
\description{
Plot function for objects returned from \code{\link{ui.pcro}}. Plots confidence intervals,
coefficients and significans assuming ignorability and the uncertainty interval under non-ignorability.
}
