% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.uiprobit.R
\name{plot.uiprobit}
\alias{plot.uiprobit}
\title{Plot of UI and CI}
\usage{
\method{plot}{uiprobit}(
  x,
  plot.all = TRUE,
  which = NA,
  intercept = FALSE,
  ylab = NULL,
  col = c("black", "red"),
  ...
)
}
\arguments{
\item{x}{An object of class uiprobit}

\item{plot.all}{If TRUE, plots all covariates.}

\item{which}{Specify which variables should be plotted by either sending in their names in a vector or a vector with their numbers (1 for the first covariate, 2 for the second etc.).To plot the intercept, set intercept as TRUE.}

\item{intercept}{If TRUE, also plots the intercept.}

\item{ylab}{Vector of names for the y-axis, default is the variable names.}

\item{col}{Vector containing the color of confidence intervals (default black) and uncertainty intervals (default red).}

\item{...}{Additional arguments, use is discouraged.}
}
\description{
Plot function for objects returned from \code{\link{ui.probit}}.
Plots confidence intervals, coefficients and significans assuming ignorability and the uncertainty interval under non-ignorability.
}
