% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ui.R
\name{print.ui}
\alias{print.ui}
\title{Print function for object of class ui}
\usage{
\method{print}{ui}(x, digits = 3, digitsci = digits, digitsui = digits, ...)
}
\arguments{
\item{x}{An object of class ui}

\item{digits}{number of digits to be printed.}

\item{digitsci}{number of digits to be printed in the confidence interval.}

\item{digitsui}{number of digits to be printed in the uncertainty interval.}

\item{...}{Additional arguments, use is discouraged.}
}
\description{
Print function for objects returned from \code{\link{ui.causal}}, \code{\link{ui.missing}}, \code{\link{ui.y0}}, \code{\link{ui.y1}}, \code{\link{ui.y0t1}}, \code{\link{ui.y1t0}}, \code{\link{ui.ace}} and \code{\link{ui.act}}.
}
