% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.uicausal.R
\name{print.uicausal}
\alias{print.uicausal}
\title{Print function for object of class uicausal}
\usage{
\method{print}{uicausal}(x, digits = 3, digitsci = digits, digitsui = digits, ...)
}
\arguments{
\item{x}{An object of returned from \code{\link{ui.causal}}}

\item{digits}{number of digits to be printed.}

\item{digitsci}{number of digits to be printed in the confidence interval.}

\item{digitsui}{number of digits to be printed in the uncertainty interval.}

\item{...}{Additional arguments, use is discouraged.}
}
\description{
Print function for object of class uicausal
}
