% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.uiols.R
\name{print.uiols}
\alias{print.uiols}
\title{Prints objects of class uiols}
\usage{
\method{print}{uiols}(x, digits = 3, digitsci = digits, digitsui = digits, ...)
}
\arguments{
\item{x}{an objects returned from \code{\link{ui.ols}}}

\item{digits}{number of digits to be printed.}

\item{digitsci}{number of digits to be printed in the confidence interval.}

\item{digitsui}{number of digits to be printed in the uncertainty interval.}

\item{...}{Additional arguments, use is discouraged.}
}
\description{
Prints objects of class uiols
}
