% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.uicausal.R
\name{profile.uicausal}
\alias{profile.uicausal}
\title{Plot of UI and CI}
\usage{
\method{profile}{uicausal}(fitted, DR = TRUE, main = "", xlab = NULL, ylab = "", ...)
}
\arguments{
\item{fitted}{An object of class uicausal}

\item{DR}{If TRUE, plots both DR if FALSE OR.}

\item{main}{Main title, default is no title.}

\item{xlab}{Title for x-axis, default is \code{expression(rho)}.}

\item{ylab}{Title for y-axis, default is the variable names.}

\item{...}{Additional arguments, use is discouraged.}
}
\description{
Plot function for objects returned from \code{\link{ui.causal}}.
Plots confidence intervals for different values of rho0=rho1=rho.
}
