% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.uiols.R
\name{profile.uiols}
\alias{profile.uiols}
\title{Plot of UI and CI}
\usage{
\method{profile}{uiols}(
  fitted,
  plot.all = TRUE,
  which = NA,
  intercept = FALSE,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{fitted}{An object of class uiols}

\item{plot.all}{If TRUE, plots all covariates.}

\item{which}{Specify which variables should be plotted by either sending in their names in a vector or a vector with their numbers (1 intercept, 2 for the first covariate etc.).}

\item{intercept}{If TRUE, also plots the intercept.}

\item{xlab}{Title for x-axis, default is \code{expression(rho)}.}

\item{ylab}{Title for y-axis, default is the variable names.}

\item{...}{Additional arguments, for instance margins.}
}
\description{
Plot function for objects returned from \code{\link{ui.ols}}.
Plots confidence intervals for different values of rho and the uncertainty interval.
}
