% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.uipcor.R
\name{profile.uipcor}
\alias{profile.uipcor}
\title{Plot of confidence intervals for partial correlation by values of the sensitivity parameter (s).}
\usage{
\method{profile}{uipcor}(x, ...)
}
\arguments{
\item{x}{An object of class uipcor}

\item{...}{Additional arguments, for instance margins.}
}
\description{
Plots confidence intervals by values of sensitivity parameter (s) for objects returned from \code{\link{ui.pcor}}.
}
\details{
Plots estimated partial correlation, lower and upper bounds of confidence intervals for different values of rho (based on on values of the sensitivity parameter(s) in \code{x$gridrho}).
Plots a 2D plot when missing data only in the outcome or the same missing in the outcome and the predictor of interest (missing data mechanisms A and B).
Plots a 3D plot when different missing data in the outcome and the predictor of interest (missing data mechanism C).
}
