% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.uiprobit.R
\name{profile.uiprobit}
\alias{profile.uiprobit}
\title{Plot of UI and CI}
\usage{
\method{profile}{uiprobit}(
  fitted,
  plot.all = TRUE,
  which = NA,
  intercept = FALSE,
  xlab = NULL,
  ylab = NULL,
  cex.lab = 2,
  mar = c(6, 6, 2, 2),
  ...
)
}
\arguments{
\item{fitted}{An object of class uiprobit}

\item{plot.all}{If TRUE, plots all covariates.}

\item{which}{Specify which variables should be plotted by either sending in their names in a vector or a vector with their numbers (1 intercept, 2 for the first covariate etc.).}

\item{intercept}{If TRUE, also plots the intercept.}

\item{xlab}{Title for x-axis, default is \code{expression(rho)}.}

\item{ylab}{Title for y-axis, default is the variable names.}

\item{cex.lab}{Size of lables.}

\item{mar}{Margin around panels in plot.}

\item{...}{Additional arguments, use is discouraged.}
}
\description{
Plot function for objects returned from \code{\link{ui.probit}}.
Plots confidence intervals for different values of rho and the uncertainty interval.
}
