% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandACT.R
\name{sandACT}
\alias{sandACT}
\title{Calculates standard error of Average causal effect on the treated}
\usage{
sandACT(
  deltasigma1,
  X,
  Xz,
  y,
  z,
  u,
  BetaOLSy0,
  phat,
  NaivEst,
  n1,
  n0,
  N,
  p,
  pz
)
}
\arguments{
\item{deltasigma1}{Coefficients.}

\item{X}{Covariate matrix outcome.}

\item{Xz}{Covariate matrix treatment.}

\item{y}{Outcome vector.}

\item{z}{Missingness indicator.}

\item{u}{Fitted values from propensity score regression.}

\item{BetaOLSy0}{Coefficients from non-treated regression}

\item{phat}{Fitted propensity scores.}

\item{NaivEst}{Naiv estimates.}

\item{n1}{Number of treated.}

\item{n0}{Number of non-treated.}

\item{N}{Total number.}

\item{p}{Number of covariates outcome regression.}

\item{pz}{Number of covariates treatment regression.}
}
\description{
This is a support function for \code{\link{ui.causal}} and calculates standard error of Average causal effect on the treated for the doubly robust estimator.
}
