% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandImpACE.R
\name{sandImpACE}
\alias{sandImpACE}
\title{Calculates standard error of Average causal effect}
\usage{
sandImpACE(X, y, z, BetaOLSy0, BetaOLSy1, NaivEst, N, p)
}
\arguments{
\item{X}{Covariate matrix.}

\item{y}{Outcome vector.}

\item{z}{missingness indicator.}

\item{BetaOLSy0}{Coefficients from non-treated regression.}

\item{BetaOLSy1}{Coefficients from treated regression.}

\item{NaivEst}{Naiv estimates.}

\item{N}{Total number.}

\item{p}{Number of covariates outcome regression.}
}
\description{
This is a support function for \code{\link{ui.causal}} and calculates standard error of Average causal effect for the regression imputation estimator.
}
