% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandImpACT.R
\name{sandImpACT}
\alias{sandImpACT}
\title{Calculates standard error of Average causal effect on the treated}
\usage{
sandImpACT(X, y, z, BetaOLSy0, NaivEst, n1, N, p)
}
\arguments{
\item{X}{Covariate matrix.}

\item{y}{Outcome vector.}

\item{z}{missingness indicator}

\item{BetaOLSy0}{Coefficients from non-treated regression}

\item{NaivEst}{Naiv estimates.}

\item{n1}{Number of treated.}

\item{N}{Total number.}

\item{p}{Number of covariates outcome regression.}
}
\description{
This is a support function for \code{\link{ui.causal}} and calculates standard error of Average causal effect on the treated for the regression imputation estimator.
}
