% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.ya.R
\name{sandwich.ya}
\alias{sandwich.ya}
\title{Calculates standard error of OR estimates of the parameters E(Y1) and E(Y0)}
\usage{
sandwich.ya(X, y, z, Beta, NaivEst, N, p, Y1 = FALSE)
}
\arguments{
\item{X}{Covariate matrix.}

\item{y}{Outcome vector.}

\item{z}{missingness indicator.}

\item{Beta}{Coefficients from the regression.}

\item{NaivEst}{Naiv estimates.}

\item{N}{Total number.}

\item{p}{Number of covariates outcome regression}

\item{Y1}{Binary variable which specifies if standard error corresponds to the estimate of average of the potential outcome under treatment.}
}
\description{
This is a support function for \code{\link{ui.causal}} and \code{\link{ui.missing}} and calculates standard error of average of the potential outcomes for the regression imputation estimator.
}
