% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaOLScor0new.R
\name{sigmaOLScor0new}
\alias{sigmaOLScor0new}
\title{Correction of OLS sigma for residual of E(Y0|X) needed for derivation of confounding bias}
\usage{
sigmaOLScor0new(sigmaOLS, u, gridrho)
}
\arguments{
\item{sigmaOLS}{Residual sd from outcome regression.}

\item{u}{Fitted values from propensity score regression.}

\item{gridrho}{Values of rho.}
}
\description{
This function is a bias correction of the estimator of residual standard deviation of E(Y0|X) under MNAR, for \code{\link{ui.causal}}.
}
