% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaOLScor1.R
\name{sigmaOLScor1}
\alias{sigmaOLScor1}
\title{Correction of OLS sigma}
\usage{
sigmaOLScor1(X, sigmaOLS, n, p, u, gridrho)
}
\arguments{
\item{X}{Covariate matrix outcome.}

\item{sigmaOLS}{Residual sd from outcome regression.}

\item{n}{Number of complete cases.}

\item{p}{Number of covariates outcome regression.}

\item{u}{Fitted values from propensity score regression.}

\item{gridrho}{Values of rho.}
}
\description{
This function is a bias correction of the residual standard deviation under MNAR, used by \code{\link{ui.causal}} and \code{\link{ui.ols}}.
}
