% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.ace.R
\name{ui.ace}
\alias{ui.ace}
\title{Uncertainty intervals for average causal effect}
\usage{
ui.ace(
  out.formula,
  y.data,
  gamma,
  t.data,
  rho0 = c(-0.3, 0.3),
  rho1 = c(-0.3, 0.3),
  sand = TRUE,
  gridn = 21,
  rho.plotrange = c(-0.5, 0.5),
  alpha = 0.05,
  sigma_correction,
  ...
)
}
\arguments{
\item{out.formula}{Formula for the outcome regression models}

\item{y.data}{data.frame containing variables needed for fitting the outcome models}

\item{gamma}{Vector of estimated coefficients from the fitted model for treatment variable}

\item{t.data}{data.frame containing variables needed for fitting the propensity score model}

\item{rho0}{Pre-specified interval for \code{rho0}.}

\item{rho1}{Pre-specified interval for \code{rho1}.}

\item{sand}{Specifies which estimator of the standard errors should be used for OR, see details in \code{\link{ui.causal}}.}

\item{gridn}{Number of fixed points within the \code{rho0} and \code{rho0} intervals for which sigma0 and sigma1 should be estimated.}

\item{rho.plotrange}{an interval larger than \code{rho0} and \code{rho1} for the plot using \code{\link{plot.ui}}.}

\item{alpha}{Default 0.05 corresponding to a confidence level of 95 for CI and UI.}

\item{sigma_correction}{A logical variable which specifies if a corrected estimation of variance of the outcome model is used to find the confounding bias. Default value is TRUE.}
}
\description{
This is a support function for \code{\link{ui.causal}} and derives uncertainty intervals for average causal effect.
}
