% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.pcor.R
\name{ui.pcor}
\alias{ui.pcor}
\title{Uncertainty intervals for partial correlation}
\usage{
ui.pcor(
  out.formula,
  data,
  rho = c(-0.3, 0.3),
  rho2 = rho,
  alpha = 0.05,
  gridn = 101,
  gridn2 = 11
)
}
\arguments{
\item{out.formula}{Regression formula where one variable of interest is the outcome and another variable of interest is the first predictor. Other predictors are variables in the adjustment set.}

\item{data}{data.frame containing the variables in the formula.}

\item{rho}{The min and the max of the sensitivity parameter.}

\item{alpha}{Significance level. Default 0.05 corresponding to a confidence level of 95% for CI and UI.}

\item{gridn}{The number of distinct points  for \code{rho} at which confidence intervals should be constructed. Default is 101.}

\item{gridn2}{The number of distinct points for \code{rho2} at which confidence intervals should be constructed for missing data mechanism C. Default is 11.}

\item{mis.formula}{Formula for missingness mechanism. If NULL the same covariates as in the outcome regression will be used.}
}
\value{
A list containing:
\item{call}{The matched call}
\item{out.formula}{Regression formula where one variable of interest is the outcome and another variable of interest is the first predictor. Other predictors are variables in the adjustment set. }
\item{out.model}{A result of a call to \code{lm} for the specified regression. }
\item{mis.formula}{Regression formula for missingness mechanism(s).}
\item{mis.model}{A result of a call to \code{glm} for probit model (s) for the missingness mechanism(s).}
\item{rho}{The range of values for the sensitivity parameter, \code{rho}, for which an uncertainty interval is constructed. Default  is (-0.3; 0.3).}
\item{gridrho}{The values of the sensitivity parameter \code{rho} (parameters \code{rho1} and \code{rho2} for missing data mechanisms C) for which confidence intervals are constructed.}
\item{pcor,rho0}{Estimated partial correlation assuming that the sensitivity parameter(s) is 0.}
\item{pcor}{Estimated partial correlation for different values of the sensitivity parameter(s) in \code{gridrho}.}
\item{ident.bound}{Bounds for the estimated identification region for partial correlation. An interval from the lowest to the largest estimated partial correlation. }
\item{pcor.se}{Standard error for different values of the sensitivity parameter(s) in \code{gridrho}.}
\item{ci.rho0}{Confidence interval for partial correlation assuming that the sensitivity parameter(s) is 0.}
\item{ci}{Confidence intervals for partial correlation for different values of the sensitivity parameter(s) in \code{gridrho}.}
\item{ui}{Uncertainty interval for partial correlation. An interval from the minimum lower bound to the maximum upper bound of estimated confidence intervals.}
}
\description{
This function is used to calculate uncertainty intervals for partial correlation between two variables adjusting for a set of other variables.
The two variables of interest might have missing data according to the three missing data mechanisms (see reference). Data for the variables in the adjustment set should be completely observed.
}
\details{
In order to visualize the results, you can use \code{\link{plot.uipcor}}. In the initial publication (see reference), the sensitivity parameter is called \gamma,  while the function denote the sensitivity paramter as \code{rho} for the correspondence with other functions in the package.
}
\examples{
library(MASS)
n <- 1000
rho <- 0.1
error <- mvrnorm(n, mu = c(0, 0, 0), Sigma = diag(c(1.16, 0.028^2 * (1 - rho^2), 1)))
X3 <- rnorm(n, mean = 67, sd = 7)
X4 <- rbinom(n, 1, prob = 0.3)
X2 <- 2.313 - 0.042 * X3 - 0.216 * X4 + error[, 1]
X1 <- 1.092 + 0.01 * X2 - 0.002 * X3 - 0.006 * X4 + 0.028 * rho * error[, 3] + error[, 2]
Z <- as.numeric(2.708 + 0.548 * X2 - 0.036 * X3 - 0.042 * X4 + error[, 3] > 0)
X1[Z == 0] <- NA
data <- data.frame(X1, X2, X3, X4)
ui <- ui.pcor(
  out.formula = X1 ~ X2 + X3 + X4,
  data = data,
  rho = c(0.1, 0.5),
  alpha = 0.05,
  gridn = 10
)
ui
}
\references{
Gorbach, T., de Luna, X. (2018). Inference for partial correlation when data are missing not at random. \emph{Statistics & Probability Letters}, 141, 82-89.
}
\author{
Tetiana Gorbach
}
