% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.y1t0.R
\name{ui.y1t0}
\alias{ui.y1t0}
\title{Uncertainty intervals for E(Y1|T=0)}
\usage{
ui.y1t0(
  out.formula,
  y.data,
  gamma,
  t.data,
  rho1 = c(-0.3, 0.3),
  sand = TRUE,
  gridn = 21,
  rho.plotrange = c(-0.5, 0.5),
  alpha = 0.05,
  sigma_correction,
  ...
)
}
\arguments{
\item{out.formula}{Formula for the outcome regression model}

\item{y.data}{data.frame containing variables needed for fitting the outcome model}

\item{gamma}{Vector of estimated coefficients from the fitted model for treatment variable}

\item{t.data}{data.frame containing variables needed for fitting the propensity score model}

\item{rho1}{Pre-specified interval for \code{rho1}.}

\item{sand}{Specifies which estimator of the standard errors should be used for OR, see details in \code{\link{ui.causal}}.}

\item{gridn}{Number of fixed points within the \code{rho1} interval for which sigma1 should be estimated.}

\item{rho.plotrange}{an interval larger than \code{rho1} for the plot using \code{\link{plot.ui}}.}

\item{alpha}{Default 0.05 corresponding to a confidence level of 95 for CI and UI.}

\item{sigma_correction}{A character variable which specifies if a corrected estimation of variance of the outcome model is used to find the confounding bias and if so whether we use the old or new correction method. It can take values 'non', 'old' or 'new'.}
}
\description{
This is a support function for \code{\link{ui.causal}} and derives the uncertainty intervals for E(Y1|T=0).
}
