% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs.hdm.R
\name{vs.hdm}
\alias{vs.hdm}
\title{#' Covariate selection using the hdm package}
\usage{
vs.hdm(Y, X, minscreen = 0)
}
\arguments{
\item{Y}{outcome vector.}

\item{X}{Covariate matrix.}
}
\description{
This is a support function for \code{\link{ui.causal}} and \code{\link{ui.missing}} and performs variable selection using logistic or linear lasso regression depending on if the output variable is binary or not.
}
