% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridrho.f.R
\name{gridrho.f}
\alias{gridrho.f}
\title{Divides an interval into a grid}
\usage{
gridrho.f(rho, gridn, rho.plotrange, plot)
}
\arguments{
\item{rho}{interval that should be divided}

\item{gridn}{number of grid points}

\item{rho.plotrange}{a larger interval of grids to be used in a plot}

\item{plot}{whether or not the larger interval of grids should be created}
}
\description{
This is a support function for \code{\link{ui.y0}}, \code{\link{ui.y1}}, \code{\link{ui.y0t1}} and \code{\link{ui.y1t0}} that divides the \code{rho} interval into a grid
}
