% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interv.p.R
\name{interv.p}
\alias{interv.p}
\title{Print interval in parantesis}
\usage{
interv.p(v, digits = 3)
}
\arguments{
\item{v}{Lower and upper bounds.}

\item{digits}{Number of decimals.}
}
\description{
This function allows you to print an interval (vector of two elements) in a parantesis single element.
}
