% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.uicausal.R
\name{plot.uicausal}
\alias{plot.uicausal}
\title{Plot ui objects}
\usage{
\method{plot}{uicausal}(
  x,
  DR = TRUE,
  main = "",
  infer = TRUE,
  lower = NULL,
  upper = NULL,
  contour = FALSE,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class ui}

\item{DR}{If TRUE the doubly robust estimator is plotted, otherwise the outcome regression estimator is plotted.}

\item{main}{Main title, default is no title.}

\item{infer}{If TRUE confidence and uncertainty intervals are provided in chosen ranges of rho, default is TRUE.}

\item{lower}{A lower bound for the parameter of interest, if available.}

\item{upper}{An upper bound for the parameter of interest, if available.}

\item{contour}{If TRUE and ACE is the parameter of interest a contour plot will be made where rho0 and rho1 are on x and y axis, respectively. Default is FALSE.}

\item{...}{Additional arguments, use is discouraged.}
}
\description{
Plot function for objects returned from \code{\link{ui.causal}}, \code{\link{ui.missing}}, \code{\link{ui.y0}}, \code{\link{ui.y1}}, \code{\link{ui.y0t1}}, \code{\link{ui.y1t0}}, \code{\link{ui.ace}} and \code{\link{ui.actT}}.
Plots confidence intervals for different values of rho and the uncertainty interval. Alternatively, a contour plot can be made in case of ACE.
}
